+++ package.json
@@ -23,7 +23,6 @@
     "express": "^4.17.3",
     "joi": "^17.6.0",
     "jsonwebtoken": "^8.5.1",
-    "moment": "^2.29.3",
     "mongodb": "^4.5.0",
     "mongoose": "^6.3.2",
     "uuid": "^8.3.2"
