--- src/routes/paymentRoutes.js
@@ -1,15 +1,0 @@
-const express = require('express');
-const router = express.Router();
-
-/* const authMiddleware = require('../middleware/authMiddleware');
-
-const paymentCtr = require('../controller/paymentController'); */
-
-
-
-
-/* router.post('/paymentIntent', authMiddleware, paymentCtr); */
-
-// router.post('/addition', validatorMiddleware(userSchemaYup), addCtrl);
-
-module.exports = router;
