--- src/schemas/operationsSchema.js
@@ -1,8 +1,0 @@
-const Joi = require('joi');
-
-module.exports = Joi.object({
-  req: {
-    value1: Joi.number().required(),
-    value2: Joi.number().required()
-  }
-}).unknown(true);
