--- src/schemas/searchParamsSchema.js
@@ -1,7 +1,0 @@
-const Joi = require('joi');
-
-module.exports = Joi.object({
-  query: Joi.object({
-    operation: Joi.string().required()
-  })
-}).unknown(true);
